Add files via upload
Screen.c | Added a few comments on the hex codes that aren't blindingly obvious with some theories on how to find our their purpose. I assume that the command function call, depending on what command is sent, the data that is sent is related to that command. i.e 0x18 has an option for 80h. refer to line 189 for some thoughts. 

Screen.h | wrote comments on thoughts on methods of testing each hex code, currently, you are able to comment out huge chunks of code, and there not be an effect on the screen, so i've written some ways I think we can test these hex codes.